livetales/backend/requirements.txt
# Core
google-adk>=1.2.0
google-genai>=1.14.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Google Cloud
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.78.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
Pillow>=11.0.0

# Dev
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
